Always suspend comet requests before resuming them.
When a container supports request suspension, we were first scheduling a
function to resume the request on a different thread and then suspending the
request. While the function was scheduled in the future, the fact that it was
scheduled before the suspend had definitively occurred meant there were thread
schedulings where the resume could occur before the suspend had executed,
leading to exceptions.

We now suspend the request *before* scheduling the resume. We schedule the
resume in a finally block to most closely track current behaviors when the
thread scheduling was correct, in which a resume would occur whether or not the
suspend threw an exception. I'm not certain about the full shape of that edge
case, so for now we take the least-changed path to dealing with it while still
addressing the core reported problem.