Add RenderSettings double rendering customization.
Rather than forcibly rendering NaN and +/-Infinity as null, we now provide a
doubleRenderer in RenderSettings. By default, we shift to the null rendering
strategy for special values. However, we also provide the old-style as-is
rendering, and a new option to throw an exception when rendering special-value
doubles for those who would prefer to not spit nulls into JSON where they
expect strictly doubles.