Add FIXME about removing finally block in next major version.
We've put the request resume scheduling in a finally block so it will run even
if the request suspend fails. This is strictly speaking incorrect---we
shouldn't attempt to process the request if we fail to suspend it altogether,
since this also means it will fail to resume. However, there could be consumer
code that relies on requests to trigger their processing even if request
suspension fails, since that is how it worked before.

As such, we mark this with a FIXME to adjust the behavior to the "correct"
behavior in the next major version, but keep the behaviorally-similar version
for the time being.