Add BoxLogging and spec.
BoxLogging provides helpers for logging boxes and then continuing to operate on
the box, alloing logging to become a fluent part of the box interaction. It
provides methods that can allow for logging empties and failures, or just
failures. It also allows logging these to different log levels, and allows for
different adapters to be applied.

For example, one could write a snippet:

  class MySnippet(id: String) extends LoggableBoxLogging {
    def render = {
      "#content" #> {
        Content.getById(id).flatMap { content =>
          content.text.asHtml // assume this returns Box[NodeSeq]
        }.infoLogEmptyBox(s"Failed to retrieve content HTML for $id")
    }
  }

When this snippet gets rendered, if the provided content lookup fails, or if
the conversion to `NodeSeq` fails, there will be an INFO level log that will
capture this failure with the provided message prefix.

BoxLogging provides handling for arbitrary-depth failure chains, as well as
providing logging for failure params, thus ensuring all of the information in a
failure will be logged. It also correctly passes any captured Throwable to the
logger methods, provided that the adapter handles them correctly.

Two adapters are provided out of the box: `LoggableBoxLogging`, which can be
mixed in as above and is based on Lift's `Loggable`, and `SLF4JBoxLogging`,
which is based on the raw SLF4J interface, and requires a protected `logger`
`var` with an SLF4J logger to be available.

Lastly, for simplicity, you can import net.liftweb.common.BoxLogging._ to have
the logging helpers available without mixing anything else into your classes.