Fix LAFuture.collect/collectAll when sub-futures fail.
Before, this would cause the overall future to hang forever. We now fail the
overall future if any contained future fails.

We also refactor the way that collect and collectAll are written to go through
a common path, which is also available for third-party use, and we add some
extra specs around collect and collectAll behavior.