Introduce PresenceBox, TryBox, and ParamTryBox.
These three intermediary types allow APIs and consumers of boxes to indicate
that they expect a particular subset of Box types. This is useful to restrict
types to what is expected, while still being able to interoperate with
APIs designed to interact with many all Box types.

For example, you can declare that a function expectes a PresenceBox to ensure
that interactors guarantee either a Full or Empty is passed. Or, you can
specify that your function returns a TryBox so callers only need to worry about
handling a Full or a Failure. Lastly, ParamTryBox allows indicating that a
given situation will only get a Full or a ParamFailure with a particular error
type.

A lot of tests and additional API conversions are required to fully bring this
vision to fruition, but this is the foundation for that work.