Make pass and $ return self.type.
This allows us to make sure that we don't lose the specific type of the Box
when doing a pass operation. This was less important when the only supertype
was `Box`, but now that we can have the more specific `PresenceBox`, `TryBox`,
and `ParamTryBox`, preserving those supertypes is more important.