This change contains two optimizations:
In appendEscapedString requires a boxing operation to transform Char into Character for the Set.contains(java.lang.Object) operation, even is the escapedChars isEmpty (the default)
The patch eliminates this check when the settings.escapeChars isEmpty using && to short circuit the Set.contains method
In Meta memoize there is an extra lookup and Option creation as both Map.contains and Map.apply are implemented using Map.get for the case where the key has already been cached (one for contains and another for get, which is creating additional Option instances).
The patch uses getOrElse to avoid the second lookup (the or else populates the cache as a side effect).