Drop attempt at flatten type specialization.
Due to the way Scala resolves method overloads with implicit parameters, I
haven't been able to find a good way of defining that a TryBox that contains a
TryBox should flatten to a TryBox, and a PresenceBox that contains a
PresenceBox should flatten to a PresenceBox. Unfortunately Scala seems to
resolve the overload of flatten based solely on class hierarchy and number of
parameters, and then fails the compile if it can't plug the proper implicits
in, rather than using the available implicit evidence to choose between
overloads. Since we rely entirely on implicit evidence to know what we can do
with the contained type of the box, there's no way (that I could find) to
express diverging return types based on the contained type :(

Instead, TryBox and PresenceBox each have their own flattenTry and
flattenPresence methods, respectively, which will only compile if their
contents match the container type. This means the caller has to choose the
appropriate one to call, but at least it ensures you *can* flatten that way if
you really want to.