Issue#1919: Asset root path should be configurable
Added 'assetPath' variable to LiftRules, with default value '/assets'. Rest of the path structure is still convention based, not configurable.

This will address the issue where the Lift template projects are storing the lazy loading gif under assets folder and the framework library expects it to reside at the root level.

On the other hand, also if the default place for the assets is in the way of the application for a reason or another, it is now easier to move somewhere else.

Added also my name to the contributors.md file, as instructed in the documentation.