Avoid surprise NPEs on param / sitemap methods
It's possible to get a surprise NPE if you try to read the currentValue
of a loc that's not the current loc instance. This is because the Loc
you're trying to read may not have a SiteMap assigned to it, which will
cause explosions in param calculation.

We now guard against this in various high-profile public API methods so
that we minimize the footprint of surprise NPEs.