Make Box compile on Scala 2.13
One superficial placeholder argument bit the 2.13 compiler insists on,
and one move of `option2Box` and `box2Option` to their own
lower-priority implicits trait that is mixed into `BoxTrait`. This
allows for proper resolution of the `box2IterableOnce` implicit as
higher priority than the `box2Option` implicit when calling `flatten` on
a `Seq[Box[T]]`.