Fix JDK 11+/Scala 2.13 build for util.Schedule
JDK 11/12 have an inaccessible `Helpers` class in
`java.util.concurrent`. In lift-utill's `Schedule` class, a reference to
`Helpers` in the context of a wildcard `java.util.concurrent._` import
is leading the compiler to try to resolve the inaccessible `Helpers` class
instead of the package-local `net.liftweb.util.Helpers`.

This commit explicitly imports `Helpers` and `Helpers.TimeSpan` to sidestep
this issue.