feat(lift-json)!: Prepare for scala 3 with type extractor (#2030)
* feat(lift-json): starting abstraction of type work

Earlier versions of lift-json depended heavily on the old scala Manifest
class that has been removed in Scala 3. The plan is to replace with with
metaprogramming in Scala 3, but to support that we first have to
encapsulate the existing Manifest logic into a separate class. This is
the starting point toward doing that.

* feat: unify on the type extractor

* fix: directly invoke fromManifest here

* fix: rip out implicit low priorty stuff

* fix: clean up non-exhaustive match