feat(lift-common)!: scala 3 support (#2031)
* chore: prepare build for scala3

* fix(Box): Replace existental type with `Any`

This was really a helper to support some explicit casting. This *may*
result in a sight change in behavior, but as best I can tell it'll still
work.

* fix(Box): explicitly invoke `ev` implicit

Changes to implicits mean we have to change how we do this. In favor of
leaving the API interface the same for now, I'm going to just driectly
invokce it.

* fix(LRU): fix indentation

* deps: update depenednecies to prep for lift3 common build

* fix(BogLogging): expose logFailure so overrides work

* fix(Conversions): don't bother with conversion here

We'll rely on the compiler finding an implicit to String in the user's
code if one is, in fact, needed

* fix(lift-common): fix BoxLoggingSpec for scala3 cross compile

* fix(lift-common): various build isssues with scala 3

* fix: scalatestplus dependency reference fix

* fix(lift-util): fix syntax errors; bypass broken tests

* chore: clean up persistence projects from build file